Saving work, confirmed db connection
